Code cleanup: remove unused & derive Default
Apply small refactors and lint fixes across multiple crates: remove unused imports, prefix unused variables with `_`, simplify Option/Result handling (e.g. unwrap_or_default, is_none/is_some_and), replace or_insert_with(...HashSet::new) with or_default, derive Default for structs/enums and use #[default] for enum variants, tighten lifetimes/signatures where appropriate, and simplify lazy/init calls. Also update pulsar_engine version in Cargo.toml to 0.1.144 (lockfile updated). These are non-functional cleanup changes to reduce warnings and improve code clarity.

+++ ui-crates/ui_log_viewer/src/gpu_info.rs
@@ -52,7 +52,7 @@ pub fn detect_primary_gpu() -> GpuInfo {
     };
 
     let info = adapter.get_info();
-    let vendor = pci_vendor_name(info.vendor as u32);
+    let vendor = pci_vendor_name(info.vendor);
     let vram_total_mb = query_vram_total_mb();
 
     // wgpu populates driver/driver_info on Vulkan; on DX12 they are often empty.
@@ -112,7 +112,7 @@ pub fn query_vram_shared_mb() -> Option<u64> {
 }
 
 fn pci_vendor_name(vendor_id: u32) -> String {
-    match vendor_id as u32 {
+    match vendor_id {
         0x10DE => "NVIDIA",
         0x1002 | 0x1022 => "AMD",
         0x8086 => "Intel",
